split tokens tests
