+++ tokens/src/lib.rs
@@ -75,6 +75,11 @@ mod imbalances;
 mod impls;
 mod mock;
 mod tests;
+mod tests_currency_adapter;
+mod tests_fungibles;
+mod tests_multicurrency;
+mod tests_events;
+
 mod weights;
 
 pub use impls::*;
