+++ tokens/src/tests.rs
@@ -1130,1374 +1130,6 @@ fn account_survive_due_to_dust_transfer_failure() {
 	});
 }
 
-// *************************************************
-// tests for MultiCurrency, MultiCurrencyExtended, MultiLockableCurrency,
-// MultiReservableCurrency traits **********************************************
-// ***
-
-#[test]
-fn multicurrency_deposit_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		assert!(!Accounts::<Runtime>::contains_key(CHARLIE, DOT));
-		assert_eq!(Tokens::free_balance(DOT, &CHARLIE), 0);
-		assert_eq!(Tokens::total_issuance(DOT), 0);
-		assert_ok!(Tokens::deposit(DOT, &CHARLIE, 10));
-		assert!(Accounts::<Runtime>::contains_key(CHARLIE, DOT));
-		assert_eq!(Tokens::free_balance(DOT, &CHARLIE), 10);
-		assert_eq!(Tokens::total_issuance(DOT), 10);
-	});
-}
-
-#[test]
-fn multicurrency_withdraw_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert!(Accounts::<Runtime>::contains_key(ALICE, DOT));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::total_issuance(DOT), 100);
-			assert_ok!(Tokens::withdraw(DOT, &ALICE, 99));
-			assert!(!Accounts::<Runtime>::contains_key(ALICE, DOT));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::total_issuance(DOT), 1);
-		});
-}
-
-#[test]
-fn multicurrency_transfer_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert!(Accounts::<Runtime>::contains_key(ALICE, DOT));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 100);
-			assert_ok!(<Tokens as MultiCurrency<_>>::transfer(DOT, &ALICE, &BOB, 99));
-			assert!(!Accounts::<Runtime>::contains_key(ALICE, DOT));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 199);
-		});
-}
-
-#[test]
-fn multicurrency_can_slash_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert!(!Tokens::can_slash(DOT, &ALICE, 101));
-			assert!(Tokens::can_slash(DOT, &ALICE, 100));
-		});
-}
-
-#[test]
-fn multicurrency_slash_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			// slashed_amount < amount
-			assert_eq!(Tokens::total_issuance(DOT), 100);
-			assert_eq!(Tokens::slash(DOT, &ALICE, 50), 0);
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_issuance(DOT), 50);
-
-			// slashed_amount == amount
-			assert_eq!(Tokens::slash(DOT, &ALICE, 51), 1);
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::total_issuance(DOT), 0);
-		});
-}
-
-#[test]
-fn multicurrency_extended_update_balance_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::update_balance(DOT, &ALICE, 50));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 150);
-			assert_eq!(Tokens::total_issuance(DOT), 250);
-
-			assert_ok!(Tokens::update_balance(DOT, &BOB, -50));
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 50);
-			assert_eq!(Tokens::total_issuance(DOT), 200);
-
-			assert_noop!(Tokens::update_balance(DOT, &BOB, -60), Error::<Runtime>::BalanceTooLow);
-		});
-}
-
-#[test]
-fn multi_lockable_currency_set_lock_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::set_lock(ID_1, DOT, &ALICE, 10));
-			assert_eq!(Tokens::accounts(&ALICE, DOT).frozen, 10);
-			assert_eq!(Tokens::accounts(&ALICE, DOT).frozen(), 10);
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 1);
-			assert_ok!(Tokens::set_lock(ID_1, DOT, &ALICE, 50));
-			assert_eq!(Tokens::accounts(&ALICE, DOT).frozen, 50);
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 1);
-			assert_ok!(Tokens::set_lock(ID_2, DOT, &ALICE, 60));
-			assert_eq!(Tokens::accounts(&ALICE, DOT).frozen, 60);
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 2);
-		});
-}
-
-#[test]
-fn multi_lockable_currency_extend_lock_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::set_lock(ID_1, DOT, &ALICE, 10));
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 1);
-			assert_eq!(Tokens::accounts(&ALICE, DOT).frozen, 10);
-			assert_ok!(Tokens::extend_lock(ID_1, DOT, &ALICE, 20));
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 1);
-			assert_eq!(Tokens::accounts(&ALICE, DOT).frozen, 20);
-			assert_ok!(Tokens::extend_lock(ID_2, DOT, &ALICE, 10));
-			assert_ok!(Tokens::extend_lock(ID_1, DOT, &ALICE, 20));
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 2);
-		});
-}
-
-#[test]
-fn multi_lockable_currency_remove_lock_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::set_lock(ID_1, DOT, &ALICE, 10));
-			assert_ok!(Tokens::set_lock(ID_2, DOT, &ALICE, 20));
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 2);
-			assert_ok!(Tokens::remove_lock(ID_2, DOT, &ALICE));
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 1);
-		});
-}
-
-#[test]
-fn multi_reservable_currency_can_reserve_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert!(Tokens::can_reserve(DOT, &ALICE, 0));
-			assert!(!Tokens::can_reserve(DOT, &ALICE, 101));
-			assert!(Tokens::can_reserve(DOT, &ALICE, 100));
-		});
-}
-
-#[test]
-fn multi_reservable_currency_slash_reserved_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::reserve(DOT, &ALICE, 50));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_issuance(DOT), 100);
-			assert_eq!(Tokens::slash_reserved(DOT, &ALICE, 0), 0);
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_issuance(DOT), 100);
-			assert_eq!(Tokens::slash_reserved(DOT, &ALICE, 100), 50);
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::total_issuance(DOT), 50);
-		});
-}
-
-#[test]
-fn multi_reservable_currency_reserve_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_noop!(Tokens::reserve(DOT, &ALICE, 101), Error::<Runtime>::BalanceTooLow);
-			assert_ok!(Tokens::reserve(DOT, &ALICE, 0));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::total_balance(DOT, &ALICE), 100);
-			assert_ok!(Tokens::reserve(DOT, &ALICE, 50));
-			System::assert_last_event(Event::Tokens(crate::Event::Reserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 50,
-			}));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_balance(DOT, &ALICE), 100);
-
-			assert_ok!(Tokens::reserve(DOT, &ALICE, 50));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::total_balance(DOT, &ALICE), 100);
-			// ensure will not trigger Endowed event
-			assert!(System::events().iter().all(|record| !matches!(
-				record.event,
-				Event::Tokens(crate::Event::Endowed {
-					currency_id: DOT,
-					who: ALICE,
-					amount: _
-				})
-			)));
-		});
-}
-
-#[test]
-fn multi_reservable_currency_unreserve_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::unreserve(DOT, &ALICE, 0), 0);
-			assert_eq!(Tokens::unreserve(DOT, &ALICE, 50), 50);
-			System::assert_last_event(Event::Tokens(crate::Event::Unreserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 0,
-			}));
-			assert_ok!(Tokens::reserve(DOT, &ALICE, 30));
-			System::assert_last_event(Event::Tokens(crate::Event::Reserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 30,
-			}));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 70);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 30);
-			assert_eq!(Tokens::unreserve(DOT, &ALICE, 15), 0);
-			System::assert_last_event(Event::Tokens(crate::Event::Unreserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 15,
-			}));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 85);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 15);
-			assert_eq!(Tokens::unreserve(DOT, &ALICE, 30), 15);
-			System::assert_last_event(Event::Tokens(crate::Event::Unreserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 15,
-			}));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 0);
-			// ensure will not trigger Endowed event
-			assert!(System::events().iter().all(|record| !matches!(
-				record.event,
-				Event::Tokens(crate::Event::Endowed {
-					currency_id: DOT,
-					who: ALICE,
-					amount: _
-				})
-			)));
-		});
-}
-
-#[test]
-fn multi_reservable_currency_repatriate_reserved_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 0);
-			assert_eq!(
-				Tokens::repatriate_reserved(DOT, &ALICE, &ALICE, 0, BalanceStatus::Free),
-				Ok(0)
-			);
-			assert_eq!(
-				Tokens::repatriate_reserved(DOT, &ALICE, &ALICE, 50, BalanceStatus::Free),
-				Ok(50)
-			);
-			// Repatriating from and to the same account, fund is `unreserved`.
-			System::assert_last_event(Event::Tokens(crate::Event::Unreserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 0,
-			}));
-
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 0);
-
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &BOB), 0);
-			assert_ok!(Tokens::reserve(DOT, &BOB, 50));
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &BOB), 50);
-			assert_eq!(
-				Tokens::repatriate_reserved(DOT, &BOB, &BOB, 60, BalanceStatus::Reserved),
-				Ok(10)
-			);
-
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &BOB), 50);
-
-			assert_eq!(
-				Tokens::repatriate_reserved(DOT, &BOB, &ALICE, 30, BalanceStatus::Reserved),
-				Ok(0)
-			);
-			System::assert_last_event(Event::Tokens(crate::Event::ReserveRepatriated {
-				currency_id: DOT,
-				from: BOB,
-				to: ALICE,
-				amount: 30,
-				status: BalanceStatus::Reserved,
-			}));
-
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 30);
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &BOB), 20);
-
-			assert_eq!(
-				Tokens::repatriate_reserved(DOT, &BOB, &ALICE, 30, BalanceStatus::Free),
-				Ok(10)
-			);
-
-			// Actual amount repatriated is 20.
-			System::assert_last_event(Event::Tokens(crate::Event::ReserveRepatriated {
-				currency_id: DOT,
-				from: BOB,
-				to: ALICE,
-				amount: 20,
-				status: BalanceStatus::Free,
-			}));
-
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 120);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 30);
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &BOB), 0);
-		});
-}
-
-#[test]
-fn slash_draw_reserved_correct() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::reserve(DOT, &ALICE, 50));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_issuance(DOT), 100);
-
-			assert_eq!(Tokens::slash(DOT, &ALICE, 80), 0);
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 20);
-			assert_eq!(Tokens::total_issuance(DOT), 20);
-
-			assert_eq!(Tokens::slash(DOT, &ALICE, 50), 30);
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::total_issuance(DOT), 0);
-		});
-}
-
-#[test]
-fn no_op_if_amount_is_zero() {
-	ExtBuilder::default().build().execute_with(|| {
-		assert_ok!(Tokens::ensure_can_withdraw(DOT, &ALICE, 0));
-		assert_ok!(Tokens::transfer(Some(ALICE).into(), BOB, DOT, 0));
-		assert_ok!(Tokens::transfer(Some(ALICE).into(), ALICE, DOT, 0));
-		assert_ok!(Tokens::deposit(DOT, &ALICE, 0));
-		assert_ok!(Tokens::withdraw(DOT, &ALICE, 0));
-		assert_eq!(Tokens::slash(DOT, &ALICE, 0), 0);
-		assert_eq!(Tokens::slash(DOT, &ALICE, 1), 1);
-		assert_ok!(Tokens::update_balance(DOT, &ALICE, 0));
-	});
-}
-
-#[test]
-fn transfer_all_trait_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (ALICE, BTC, 200)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::free_balance(BTC, &ALICE), 200);
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 0);
-
-			assert_ok!(<Tokens as TransferAll<AccountId>>::transfer_all(&ALICE, &BOB));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::free_balance(BTC, &ALICE), 0);
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 100);
-			assert_eq!(Tokens::free_balance(BTC, &BOB), 200);
-
-			assert_ok!(Tokens::reserve(DOT, &BOB, 1));
-			assert_ok!(<Tokens as TransferAll<AccountId>>::transfer_all(&BOB, &ALICE));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 99);
-			assert_eq!(Tokens::free_balance(BTC, &ALICE), 200);
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 0);
-			assert_eq!(Tokens::free_balance(BTC, &BOB), 0);
-		});
-}
-
-#[test]
-fn named_multi_reservable_currency_slash_reserved_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &ALICE, 50));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_issuance(DOT), 100);
-			assert_eq!(Tokens::slash_reserved_named(&RID_1, DOT, &ALICE, 0), 0);
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_issuance(DOT), 100);
-			assert_eq!(Tokens::slash_reserved_named(&RID_1, DOT, &ALICE, 100), 50);
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 0);
-			assert_eq!(Tokens::total_issuance(DOT), 50);
-		});
-}
-
-#[test]
-fn named_multi_reservable_currency_reserve_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_noop!(
-				Tokens::reserve_named(&RID_1, DOT, &ALICE, 101),
-				Error::<Runtime>::BalanceTooLow
-			);
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &ALICE, 0));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 0);
-			assert_eq!(Tokens::total_balance(DOT, &ALICE), 100);
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &ALICE, 50));
-			System::assert_last_event(Event::Tokens(crate::Event::Reserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 50,
-			}));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_balance(DOT, &ALICE), 100);
-
-			assert_ok!(Tokens::reserve_named(&RID_2, DOT, &ALICE, 50));
-			System::assert_last_event(Event::Tokens(crate::Event::Reserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 50,
-			}));
-
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_2, DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_balance(DOT, &ALICE), 100);
-
-			// ensure will not trigger Endowed event
-			assert!(System::events().iter().all(|record| !matches!(
-				record.event,
-				Event::Tokens(crate::Event::Endowed {
-					currency_id: DOT,
-					who: ALICE,
-					amount: _
-				})
-			)));
-		});
-}
-
-#[test]
-fn named_multi_reservable_currency_unreserve_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 0);
-			assert_eq!(Tokens::unreserve_named(&RID_1, DOT, &ALICE, 0), 0);
-
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &ALICE, 30));
-			System::assert_last_event(Event::Tokens(crate::Event::Reserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 30,
-			}));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 70);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 30);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 30);
-
-			assert_ok!(Tokens::reserve_named(&RID_2, DOT, &ALICE, 30));
-			System::assert_last_event(Event::Tokens(crate::Event::Reserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 30,
-			}));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 40);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 60);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 30);
-			assert_eq!(Tokens::reserved_balance_named(&RID_2, DOT, &ALICE), 30);
-
-			assert_eq!(Tokens::unreserve_named(&RID_1, DOT, &ALICE, 30), 0);
-			System::assert_last_event(Event::Tokens(crate::Event::Unreserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 30,
-			}));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 70);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 30);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 0);
-			assert_eq!(Tokens::reserved_balance_named(&RID_2, DOT, &ALICE), 30);
-
-			assert_eq!(Tokens::unreserve_named(&RID_2, DOT, &ALICE, 30), 0);
-			System::assert_last_event(Event::Tokens(crate::Event::Unreserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 30,
-			}));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 0);
-			assert_eq!(Tokens::reserved_balance_named(&RID_2, DOT, &ALICE), 0);
-			// ensure will not trigger Endowed event
-			assert!(System::events().iter().all(|record| !matches!(
-				record.event,
-				Event::Tokens(crate::Event::Endowed {
-					currency_id: DOT,
-					who: ALICE,
-					amount: _
-				})
-			)));
-		});
-}
-
-#[test]
-fn named_multi_reservable_currency_repatriate_reserved_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 0);
-			assert_eq!(
-				Tokens::repatriate_reserved_named(&RID_1, DOT, &ALICE, &ALICE, 0, BalanceStatus::Free),
-				Ok(0)
-			);
-			assert_eq!(
-				Tokens::repatriate_reserved_named(&RID_1, DOT, &ALICE, &ALICE, 50, BalanceStatus::Free),
-				Ok(50)
-			);
-
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 0);
-
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 100);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &BOB), 0);
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &BOB, 50));
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &BOB), 50);
-			assert_eq!(
-				Tokens::repatriate_reserved_named(&RID_1, DOT, &BOB, &BOB, 60, BalanceStatus::Reserved),
-				Ok(10)
-			);
-
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &BOB), 50);
-
-			assert_eq!(
-				Tokens::repatriate_reserved_named(&RID_1, DOT, &BOB, &ALICE, 30, BalanceStatus::Reserved),
-				Ok(0)
-			);
-			System::assert_last_event(Event::Tokens(crate::Event::ReserveRepatriated {
-				currency_id: DOT,
-				from: BOB,
-				to: ALICE,
-				amount: 30,
-				status: BalanceStatus::Reserved,
-			}));
-
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 30);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 30);
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &BOB), 20);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &BOB), 20);
-
-			assert_eq!(
-				Tokens::repatriate_reserved_named(&RID_1, DOT, &BOB, &ALICE, 30, BalanceStatus::Free),
-				Ok(10)
-			);
-
-			// Actual amount repatriated is 20.
-			System::assert_last_event(Event::Tokens(crate::Event::ReserveRepatriated {
-				currency_id: DOT,
-				from: BOB,
-				to: ALICE,
-				amount: 20,
-				status: BalanceStatus::Free,
-			}));
-
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 120);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 30);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 30);
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 50);
-			assert_eq!(Tokens::reserved_balance(DOT, &BOB), 0);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &BOB), 0);
-		});
-}
-
-#[test]
-fn slashed_reserved_named_works() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &ALICE, 50));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_issuance(DOT), 100);
-
-			assert_eq!(Tokens::slash_reserved_named(&RID_1, DOT, &ALICE, 20), 0);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 30);
-			assert_eq!(Tokens::total_issuance(DOT), 80);
-
-			assert_eq!(Tokens::slash_reserved_named(&RID_1, DOT, &ALICE, 40), 10);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 0);
-			assert_eq!(Tokens::reserved_balance(DOT, &ALICE), 0);
-			assert_eq!(Tokens::total_issuance(DOT), 50);
-		});
-}
-
-#[test]
-fn named_multi_reservable_ensure_named_reserved_works() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &ALICE, 50));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 50);
-			assert_eq!(Tokens::total_issuance(DOT), 100);
-
-			assert_ok!(Tokens::ensure_reserved_named(&RID_1, DOT, &ALICE, 20));
-			assert_ok!(Tokens::ensure_reserved_named(&RID_1, DOT, &ALICE, 70));
-
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 30);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 70);
-		});
-}
-
-#[test]
-fn named_multi_reservable_unreserve_all_named() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &ALICE, 50));
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &ALICE, 20));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 30);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 70);
-
-			let value = Tokens::unreserve_all_named(&RID_1, DOT, &ALICE);
-			assert!(value == 70);
-
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 0);
-		});
-}
-
-#[test]
-fn named_multi_reservable_slash_all_reserved_named() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &ALICE, 50));
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 50);
-
-			let value = Tokens::slash_all_reserved_named(&RID_1, DOT, &ALICE);
-			assert!(value == 0);
-
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 50);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 0);
-		});
-}
-
-#[test]
-fn named_multi_reservable_repatriate_all_reserved_named_works() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(Tokens::free_balance(DOT, &ALICE), 100);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &ALICE), 0);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &BOB), 0);
-			assert_ok!(Tokens::reserve_named(&RID_1, DOT, &ALICE, 50));
-
-			assert_ok!(Tokens::repatriate_all_reserved_named(
-				&RID_1,
-				DOT,
-				&ALICE,
-				&BOB,
-				BalanceStatus::Reserved
-			));
-
-			assert_eq!(Tokens::free_balance(DOT, &BOB), 100);
-			assert_eq!(Tokens::reserved_balance_named(&RID_1, DOT, &BOB), 50);
-
-			System::assert_last_event(Event::Tokens(crate::Event::ReserveRepatriated {
-				currency_id: DOT,
-				from: ALICE,
-				to: BOB,
-				amount: 50,
-				status: BalanceStatus::Reserved,
-			}));
-		});
-}
-
-// *************************************************
-// tests for CurrencyAdapter
-// *************************************************
-
-#[test]
-fn currency_adapter_ensure_currency_adapter_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(Tokens::total_issuance(DOT), 100);
-			assert_eq!(Tokens::total_balance(DOT, &TREASURY_ACCOUNT), 100);
-			assert_eq!(Tokens::reserved_balance(DOT, &TREASURY_ACCOUNT), 0);
-			assert_eq!(Tokens::free_balance(DOT, &TREASURY_ACCOUNT), 100);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_balance(&TREASURY_ACCOUNT),
-				100
-			);
-			assert!(<Runtime as pallet_elections_phragmen::Config>::Currency::can_slash(
-				&TREASURY_ACCOUNT,
-				10
-			));
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
-				100
-			);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::minimum_balance(),
-				2
-			);
-			assert!(<Runtime as pallet_elections_phragmen::Config>::Currency::can_reserve(
-				&TREASURY_ACCOUNT,
-				5
-			));
-
-			// burn
-			let imbalance = <Runtime as pallet_elections_phragmen::Config>::Currency::burn(10);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
-				90
-			);
-			drop(imbalance);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
-				100
-			);
-
-			// issue
-			let imbalance = <Runtime as pallet_elections_phragmen::Config>::Currency::issue(20);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
-				120
-			);
-			drop(imbalance);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
-				100
-			);
-
-			// transfer
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
-				100
-			);
-			assert_ok!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::ensure_can_withdraw(
-					&TREASURY_ACCOUNT,
-					10,
-					WithdrawReasons::TRANSFER,
-					0
-				)
-			);
-			assert_ok!(<Runtime as pallet_elections_phragmen::Config>::Currency::transfer(
-				&TREASURY_ACCOUNT,
-				&ALICE,
-				11,
-				ExistenceRequirement::KeepAlive
-			));
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
-				89
-			);
-
-			// deposit
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
-				100
-			);
-			let imbalance = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 11);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
-				100
-			);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
-				100
-			);
-			drop(imbalance);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
-				100
-			);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
-				111
-			);
-
-			// withdraw
-			let imbalance = <Runtime as pallet_elections_phragmen::Config>::Currency::withdraw(
-				&TREASURY_ACCOUNT,
-				10,
-				WithdrawReasons::TRANSFER,
-				ExistenceRequirement::KeepAlive,
-			);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
-				90
-			);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
-				111
-			);
-			drop(imbalance);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
-				90
-			);
-			assert_eq!(
-				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
-				101
-			);
-		});
-}
-
-#[test]
-fn currency_adapter_burn_must_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			let init_total_issuance = TreasuryCurrencyAdapter::total_issuance();
-			let imbalance = TreasuryCurrencyAdapter::burn(10);
-			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), init_total_issuance - 10);
-			drop(imbalance);
-			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), init_total_issuance);
-		});
-}
-
-#[test]
-fn currency_adapter_reserving_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
-
-		assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 111);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 111);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 0);
-
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 69));
-
-		assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 111);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 42);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 69);
-	});
-}
-
-#[test]
-fn currency_adapter_balance_transfer_when_reserved_should_not_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 69));
-		assert_noop!(
-			TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 69, ExistenceRequirement::AllowDeath),
-			Error::<Runtime>::BalanceTooLow,
-		);
-	});
-}
-
-#[test]
-fn currency_adapter_deducting_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 69));
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 42);
-	});
-}
-
-#[test]
-fn currency_adapter_refunding_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 42);
-		Tokens::set_reserved_balance(DOT, &TREASURY_ACCOUNT, 69);
-		TreasuryCurrencyAdapter::unreserve(&TREASURY_ACCOUNT, 69);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 111);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 0);
-	});
-}
-
-#[test]
-fn currency_adapter_slashing_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 69));
-		assert!(TreasuryCurrencyAdapter::slash(&TREASURY_ACCOUNT, 69).1.is_zero());
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 42);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 42);
-	});
-}
-
-#[test]
-fn currency_adapter_slashing_incomplete_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 42);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 42);
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 21));
-		assert_eq!(TreasuryCurrencyAdapter::slash(&TREASURY_ACCOUNT, 69).1, 27);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 0);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 0);
-	});
-}
-
-#[test]
-fn currency_adapter_basic_locking_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 100);
-			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 91, WithdrawReasons::all());
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 10, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-		});
-}
-
-#[test]
-fn currency_adapter_partial_locking_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 5, WithdrawReasons::all());
-			assert_ok!(TreasuryCurrencyAdapter::transfer(
-				&TREASURY_ACCOUNT,
-				&ALICE,
-				2,
-				ExistenceRequirement::AllowDeath
-			));
-		});
-}
-
-#[test]
-fn currency_adapter_lock_removal_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, u64::max_value(), WithdrawReasons::all());
-			TreasuryCurrencyAdapter::remove_lock(ID_1, &TREASURY_ACCOUNT);
-			assert_ok!(TreasuryCurrencyAdapter::transfer(
-				&TREASURY_ACCOUNT,
-				&ALICE,
-				2,
-				ExistenceRequirement::AllowDeath
-			));
-		});
-}
-
-#[test]
-fn currency_adapter_lock_replacement_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, u64::max_value(), WithdrawReasons::all());
-			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 5, WithdrawReasons::all());
-			assert_ok!(TreasuryCurrencyAdapter::transfer(
-				&TREASURY_ACCOUNT,
-				&ALICE,
-				2,
-				ExistenceRequirement::AllowDeath
-			));
-		});
-}
-
-#[test]
-fn currency_adapter_double_locking_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 5, WithdrawReasons::empty());
-			TreasuryCurrencyAdapter::set_lock(ID_2, &TREASURY_ACCOUNT, 5, WithdrawReasons::all());
-			assert_ok!(TreasuryCurrencyAdapter::transfer(
-				&TREASURY_ACCOUNT,
-				&ALICE,
-				2,
-				ExistenceRequirement::AllowDeath
-			));
-		});
-}
-
-#[test]
-fn currency_adapter_combination_locking_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			// withdrawReasons not work
-			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, u64::max_value(), WithdrawReasons::empty());
-			TreasuryCurrencyAdapter::set_lock(ID_2, &TREASURY_ACCOUNT, 0, WithdrawReasons::all());
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 2, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-		});
-}
-
-#[test]
-fn currency_adapter_lock_value_extension_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 100, WithdrawReasons::all());
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 6, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 2, WithdrawReasons::all());
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 6, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 8, WithdrawReasons::all());
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 3, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-		});
-}
-
-#[test]
-fn currency_adapter_lock_block_number_extension_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 200, WithdrawReasons::all());
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 6, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 90, WithdrawReasons::all());
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 6, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-			System::set_block_number(2);
-			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 90, WithdrawReasons::all());
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 3, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-		});
-}
-
-#[test]
-fn currency_adapter_lock_reasons_extension_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 90, WithdrawReasons::TRANSFER);
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 11, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 90, WithdrawReasons::empty());
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 11, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 90, WithdrawReasons::RESERVE);
-			assert_noop!(
-				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 11, ExistenceRequirement::AllowDeath),
-				Error::<Runtime>::LiquidityRestrictions
-			);
-		});
-}
-
-#[test]
-fn currency_adapter_deposit_creating_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 100);
-			assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 0);
-			let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 2);
-			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 102);
-			assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 2);
-
-			assert_eq!(TreasuryCurrencyAdapter::total_balance(&ALICE), 100);
-			let _ = TreasuryCurrencyAdapter::deposit_creating(&ALICE, 1);
-			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 103);
-			assert_eq!(TreasuryCurrencyAdapter::total_balance(&ALICE), 101);
-		});
-}
-
-#[test]
-fn currency_adapter_deposit_into_existing_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 0);
-			assert_noop!(
-				TreasuryCurrencyAdapter::deposit_into_existing(&TREASURY_ACCOUNT, 10).map(drop),
-				Error::<Runtime>::DeadAccount,
-			);
-
-			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 100);
-			assert_eq!(TreasuryCurrencyAdapter::total_balance(&ALICE), 100);
-			assert_ok!(TreasuryCurrencyAdapter::deposit_into_existing(&ALICE, 10).map(drop));
-			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 110);
-			assert_eq!(TreasuryCurrencyAdapter::total_balance(&ALICE), 110);
-		});
-}
-
-#[test]
-fn currency_adapter_reward_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 100);
-			assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 100);
-			assert_ok!(TreasuryCurrencyAdapter::deposit_into_existing(&TREASURY_ACCOUNT, 10).map(drop));
-			assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 110);
-			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 110);
-		});
-}
-
-#[test]
-fn currency_adapter_slashing_reserved_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 111);
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 111));
-		assert_eq!(TreasuryCurrencyAdapter::slash_reserved(&TREASURY_ACCOUNT, 42).1, 0);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 69);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 69);
-	});
-}
-
-#[test]
-fn currency_adapter_slashing_incomplete_reserved_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 111);
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 42));
-		assert_eq!(TreasuryCurrencyAdapter::slash_reserved(&TREASURY_ACCOUNT, 69).1, 27);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 69);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 0);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 69);
-	});
-}
-
-#[test]
-fn currency_adapter_repatriating_reserved_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 110);
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&ALICE, 2);
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 110));
-		assert_ok!(
-			TreasuryCurrencyAdapter::repatriate_reserved(&TREASURY_ACCOUNT, &ALICE, 41, Status::Free),
-			0
-		);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 69);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&ALICE), 0);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&ALICE), 43);
-	});
-}
-
-#[test]
-fn currency_adapter_transferring_reserved_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 110);
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&ALICE, 2);
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 110));
-		assert_ok!(
-			TreasuryCurrencyAdapter::repatriate_reserved(&TREASURY_ACCOUNT, &ALICE, 41, Status::Reserved),
-			0
-		);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 69);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&ALICE), 41);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&ALICE), 2);
-	});
-}
-
-#[test]
-fn currency_adapter_transferring_reserved_balance_to_nonexistent_should_fail() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 111));
-		assert_ok!(TreasuryCurrencyAdapter::repatriate_reserved(
-			&TREASURY_ACCOUNT,
-			&ALICE,
-			42,
-			Status::Free
-		));
-	});
-}
-
-#[test]
-fn currency_adapter_transferring_incomplete_reserved_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 110);
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&ALICE, 2);
-		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 41));
-		assert_ok!(
-			TreasuryCurrencyAdapter::repatriate_reserved(&TREASURY_ACCOUNT, &ALICE, 69, Status::Free),
-			28
-		);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 0);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 69);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&ALICE), 0);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&ALICE), 43);
-	});
-}
-
-#[test]
-fn currency_adapter_transferring_too_high_value_should_not_panic() {
-	ExtBuilder::default().build().execute_with(|| {
-		TreasuryCurrencyAdapter::make_free_balance_be(&TREASURY_ACCOUNT, u64::max_value());
-		TreasuryCurrencyAdapter::make_free_balance_be(&ALICE, 2);
-
-		assert_noop!(
-			TreasuryCurrencyAdapter::transfer(
-				&TREASURY_ACCOUNT,
-				&ALICE,
-				u64::max_value(),
-				ExistenceRequirement::AllowDeath
-			),
-			ArithmeticError::Overflow,
-		);
-
-		assert_eq!(
-			TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT),
-			u64::max_value()
-		);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&ALICE), 2);
-	});
-}
-
-#[test]
-fn exceeding_max_locks_should_fail() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(Tokens::set_lock(ID_1, DOT, &ALICE, 10));
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 1);
-			assert_ok!(Tokens::set_lock(ID_2, DOT, &ALICE, 10));
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 2);
-			assert_noop!(
-				Tokens::set_lock(ID_3, DOT, &ALICE, 10),
-				Error::<Runtime>::MaxLocksExceeded
-			);
-			assert_eq!(Tokens::locks(ALICE, DOT).len(), 2);
-		});
-}
-
-#[test]
-fn currency_adapter_slashing_named_reserved_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 111);
-		assert_ok!(TreasuryCurrencyAdapter::reserve_named(&RID_1, &TREASURY_ACCOUNT, 111));
-		assert_eq!(
-			TreasuryCurrencyAdapter::slash_reserved_named(&RID_1, &TREASURY_ACCOUNT, 42).1,
-			0
-		);
-		assert_eq!(
-			TreasuryCurrencyAdapter::reserved_balance_named(&RID_1, &TREASURY_ACCOUNT),
-			69
-		);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 69);
-	});
-}
-
-#[test]
-fn currency_adapter_named_slashing_incomplete_reserved_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 111);
-		assert_ok!(TreasuryCurrencyAdapter::reserve_named(&RID_1, &TREASURY_ACCOUNT, 42));
-		assert_eq!(
-			TreasuryCurrencyAdapter::slash_reserved_named(&RID_1, &TREASURY_ACCOUNT, 69).1,
-			27
-		);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 69);
-		assert_eq!(
-			TreasuryCurrencyAdapter::reserved_balance_named(&RID_1, &TREASURY_ACCOUNT),
-			0
-		);
-		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 69);
-	});
-}
-
-#[test]
-fn currency_adapter_repatriating_named_reserved_balance_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 110);
-		let _ = TreasuryCurrencyAdapter::deposit_creating(&ALICE, 2);
-		assert_ok!(TreasuryCurrencyAdapter::reserve_named(&RID_1, &TREASURY_ACCOUNT, 110));
-		assert_ok!(
-			TreasuryCurrencyAdapter::repatriate_reserved_named(&RID_1, &TREASURY_ACCOUNT, &ALICE, 41, Status::Free),
-			0
-		);
-		assert_eq!(
-			TreasuryCurrencyAdapter::reserved_balance_named(&RID_1, &TREASURY_ACCOUNT),
-			69
-		);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
-		assert_eq!(TreasuryCurrencyAdapter::reserved_balance_named(&RID_1, &ALICE), 0);
-		assert_eq!(TreasuryCurrencyAdapter::free_balance(&ALICE), 43);
-	});
-}
-
 #[test]
 fn exceeding_max_reserves_should_fail() {
 	ExtBuilder::default()
@@ -2511,604 +1143,5 @@ fn exceeding_max_reserves_should_fail() {
 				Tokens::reserve_named(&id_3, DOT, &ALICE, 10),
 				Error::<Runtime>::TooManyReserves
 			);
-		});
-}
-
-// *************************************************
-// tests for fungibles traits
-// *************************************************
-
-#[test]
-fn fungibles_inspect_trait_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::total_issuance(DOT), 100);
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::minimum_balance(DOT), 2);
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &ALICE), 100);
-			assert_eq!(
-				<Tokens as fungibles::Inspect<_>>::reducible_balance(DOT, &ALICE, true),
-				98
-			);
-			assert_ok!(<Tokens as fungibles::Inspect<_>>::can_deposit(DOT, &ALICE, 1).into_result());
-			assert_ok!(<Tokens as fungibles::Inspect<_>>::can_withdraw(DOT, &ALICE, 1).into_result());
-		});
-}
-
-#[test]
-fn fungibles_mutate_trait_should_work() {
-	ExtBuilder::default().build().execute_with(|| {
-		assert_ok!(<Tokens as fungibles::Mutate<_>>::mint_into(DOT, &ALICE, 10));
-		assert_eq!(<Tokens as fungibles::Mutate<_>>::burn_from(DOT, &ALICE, 8), Ok(8));
-	});
-}
-
-#[test]
-fn fungibles_transfer_trait_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &ALICE), 100);
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &BOB), 100);
-			assert_ok!(<Tokens as fungibles::Transfer<_>>::transfer(
-				DOT, &ALICE, &BOB, 10, true
-			));
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &ALICE), 90);
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &BOB), 110);
-		});
-}
-
-#[test]
-fn fungibles_unbalanced_trait_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &ALICE), 100);
-			assert_ok!(<Tokens as fungibles::Unbalanced<_>>::set_balance(DOT, &ALICE, 10));
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &ALICE), 10);
-
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::total_issuance(DOT), 100);
-			<Tokens as fungibles::Unbalanced<_>>::set_total_issuance(DOT, 10);
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::total_issuance(DOT), 10);
-		});
-}
-
-#[test]
-fn fungibles_inspect_hold_trait_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &ALICE), 0);
-			assert!(<Tokens as fungibles::InspectHold<_>>::can_hold(DOT, &ALICE, 50));
-			assert!(!<Tokens as fungibles::InspectHold<_>>::can_hold(DOT, &ALICE, 100));
-		});
-}
-
-#[test]
-fn fungibles_mutate_hold_trait_should_work() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_noop!(
-				<Tokens as fungibles::MutateHold<_>>::hold(DOT, &ALICE, 200),
-				Error::<Runtime>::BalanceTooLow
-			);
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &ALICE), 0);
-			assert_ok!(<Tokens as fungibles::MutateHold<_>>::hold(DOT, &ALICE, 100));
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &ALICE), 100);
-			assert_eq!(
-				<Tokens as fungibles::MutateHold<_>>::release(DOT, &ALICE, 40, false),
-				Ok(40)
-			);
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &ALICE), 60);
-
-			// exceed hold amount when not in best_effort
-			assert_noop!(
-				<Tokens as fungibles::MutateHold<_>>::release(DOT, &ALICE, 61, false),
-				Error::<Runtime>::BalanceTooLow
-			);
-
-			// exceed hold amount when in best_effort
-			assert_eq!(
-				<Tokens as fungibles::MutateHold<_>>::release(DOT, &ALICE, 61, true),
-				Ok(60)
-			);
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &ALICE), 0);
-
-			assert_ok!(<Tokens as fungibles::MutateHold<_>>::hold(DOT, &ALICE, 70));
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &ALICE), 70);
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &BOB), 100);
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &BOB), 0);
-			assert_eq!(
-				<Tokens as fungibles::MutateHold<_>>::transfer_held(DOT, &ALICE, &BOB, 5, false, false),
-				Ok(5)
-			);
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &ALICE), 65);
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &BOB), 105);
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &BOB), 0);
-			assert_eq!(
-				<Tokens as fungibles::MutateHold<_>>::transfer_held(DOT, &ALICE, &BOB, 5, false, true),
-				Ok(5)
-			);
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &ALICE), 60);
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &BOB), 110);
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &BOB), 5);
-
-			// exceed hold amount when not in best_effort
-			assert_noop!(
-				<Tokens as fungibles::MutateHold<_>>::transfer_held(DOT, &ALICE, &BOB, 61, false, true),
-				Error::<Runtime>::BalanceTooLow
-			);
-
-			// exceed hold amount when in best_effort
-			assert_eq!(
-				<Tokens as fungibles::MutateHold<_>>::transfer_held(DOT, &ALICE, &BOB, 61, true, true),
-				Ok(60)
-			);
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &ALICE), 0);
-			assert_eq!(<Tokens as fungibles::Inspect<_>>::balance(DOT, &BOB), 170);
-			assert_eq!(<Tokens as fungibles::InspectHold<_>>::balance_on_hold(DOT, &BOB), 65);
-		});
-}
-
-#[test]
-fn fungibles_inspect_convert_should_work() {
-	pub struct ConvertBalanceTest;
-	impl ConvertBalance<Balance, Balance> for ConvertBalanceTest {
-		type AssetId = CurrencyId;
-		fn convert_balance(balance: Balance, _asset_id: CurrencyId) -> Balance {
-			balance * 100
-		}
-
-		fn convert_balance_back(balance: Balance, _asset_id: CurrencyId) -> Balance {
-			balance / 100
-		}
-	}
-
-	pub struct IsLiquidToken;
-	impl Contains<CurrencyId> for IsLiquidToken {
-		fn contains(currency_id: &CurrencyId) -> bool {
-			matches!(currency_id, &DOT)
-		}
-	}
-
-	pub struct GetCurrencyId;
-	impl Get<CurrencyId> for GetCurrencyId {
-		fn get() -> CurrencyId {
-			DOT
-		}
-	}
-
-	type RebaseTokens = Combiner<
-		AccountId,
-		IsLiquidToken,
-		Mapper<AccountId, Tokens, ConvertBalanceTest, Balance, GetCurrencyId>,
-		Tokens,
-	>;
-
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(
-				<RebaseTokens as fungibles::Inspect<AccountId>>::balance(DOT, &ALICE),
-				10000
-			);
-			assert_eq!(
-				<RebaseTokens as fungibles::Inspect<AccountId>>::total_issuance(DOT),
-				20000
-			);
-		});
-}
-
-#[test]
-fn fungibles_transfers_convert_should_work() {
-	pub struct ConvertBalanceTest;
-	impl ConvertBalance<Balance, Balance> for ConvertBalanceTest {
-		type AssetId = CurrencyId;
-		fn convert_balance(balance: Balance, _asset_id: CurrencyId) -> Balance {
-			balance * 100
-		}
-
-		fn convert_balance_back(balance: Balance, _asset_id: CurrencyId) -> Balance {
-			balance / 100
-		}
-	}
-
-	pub struct IsLiquidToken;
-	impl Contains<CurrencyId> for IsLiquidToken {
-		fn contains(currency_id: &CurrencyId) -> bool {
-			matches!(currency_id, &DOT)
-		}
-	}
-
-	pub struct GetCurrencyId;
-	impl Get<CurrencyId> for GetCurrencyId {
-		fn get() -> CurrencyId {
-			DOT
-		}
-	}
-
-	type RebaseTokens = Combiner<
-		AccountId,
-		IsLiquidToken,
-		Mapper<AccountId, Tokens, ConvertBalanceTest, Balance, GetCurrencyId>,
-		Tokens,
-	>;
-
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 300), (BOB, DOT, 200)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(<RebaseTokens as fungibles::Transfer<AccountId>>::transfer(
-				DOT, &ALICE, &BOB, 10000, true
-			));
-			assert_eq!(
-				<RebaseTokens as fungibles::Inspect<AccountId>>::balance(DOT, &ALICE),
-				20000
-			);
-			assert_eq!(
-				<RebaseTokens as fungibles::Inspect<AccountId>>::balance(DOT, &BOB),
-				30000
-			);
-		});
-}
-
-#[test]
-fn fungibles_mutate_convert_should_work() {
-	pub struct ConvertBalanceTest;
-	impl ConvertBalance<Balance, Balance> for ConvertBalanceTest {
-		type AssetId = CurrencyId;
-		fn convert_balance(balance: Balance, _asset_id: CurrencyId) -> Balance {
-			balance * 100
-		}
-
-		fn convert_balance_back(balance: Balance, _asset_id: CurrencyId) -> Balance {
-			balance / 100
-		}
-	}
-
-	pub struct IsLiquidToken;
-	impl Contains<CurrencyId> for IsLiquidToken {
-		fn contains(currency_id: &CurrencyId) -> bool {
-			matches!(currency_id, &DOT)
-		}
-	}
-
-	pub struct GetCurrencyId;
-	impl Get<CurrencyId> for GetCurrencyId {
-		fn get() -> CurrencyId {
-			DOT
-		}
-	}
-
-	type RebaseTokens = Combiner<
-		AccountId,
-		IsLiquidToken,
-		Mapper<AccountId, Tokens, ConvertBalanceTest, Balance, GetCurrencyId>,
-		Tokens,
-	>;
-
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 300), (BOB, DOT, 200)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(<RebaseTokens as fungibles::Mutate<AccountId>>::mint_into(
-				DOT, &ALICE, 10000
-			));
-			assert_ok!(<RebaseTokens as fungibles::Mutate<AccountId>>::burn_from(
-				DOT, &BOB, 10000
-			));
-			assert_eq!(
-				<RebaseTokens as fungibles::Inspect<AccountId>>::balance(DOT, &ALICE),
-				40000
-			);
-			assert_eq!(
-				<RebaseTokens as fungibles::Inspect<AccountId>>::balance(DOT, &BOB),
-				10000
-			);
-		});
-}
-
-// *************************************************
-// tests for Event Deposit
-// *************************************************
-#[test]
-fn pallet_multicurrency_deposit_events() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(<Tokens as MultiCurrency<AccountId>>::transfer(DOT, &ALICE, &BOB, 10));
-			System::assert_last_event(Event::Tokens(crate::Event::Transfer {
-				currency_id: DOT,
-				from: ALICE,
-				to: BOB,
-				amount: 10,
-			}));
-
-			assert_ok!(<Tokens as MultiCurrency<AccountId>>::deposit(DOT, &ALICE, 10));
-			System::assert_last_event(Event::Tokens(crate::Event::Deposited {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 10,
-			}));
-
-			assert_ok!(<Tokens as MultiCurrency<AccountId>>::withdraw(DOT, &ALICE, 10));
-			System::assert_last_event(Event::Tokens(crate::Event::Withdrawn {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 10,
-			}));
-
-			assert_ok!(<Tokens as MultiReservableCurrency<AccountId>>::reserve(DOT, &ALICE, 50));
-			assert_eq!(<Tokens as MultiCurrency<AccountId>>::slash(DOT, &ALICE, 60), 0);
-			System::assert_last_event(Event::Tokens(crate::Event::Slashed {
-				currency_id: DOT,
-				who: ALICE,
-				free_amount: 40,
-				reserved_amount: 20,
-			}));
-		});
-}
-
-#[test]
-fn pallet_multicurrency_extended_deposit_events() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(<Tokens as MultiCurrencyExtended<AccountId>>::update_balance(
-				DOT, &ALICE, 500
-			));
-			System::assert_last_event(Event::Tokens(crate::Event::Deposited {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 500,
-			}));
-			assert_ok!(<Tokens as MultiCurrencyExtended<AccountId>>::update_balance(
-				DOT, &ALICE, -500
-			));
-			System::assert_last_event(Event::Tokens(crate::Event::Withdrawn {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 500,
-			}));
-		});
-}
-
-#[test]
-fn pallet_multi_lockable_currency_deposit_events() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(<Tokens as MultiLockableCurrency<AccountId>>::set_lock(
-				[0u8; 8], DOT, &ALICE, 10
-			));
-			System::assert_last_event(Event::Tokens(crate::Event::LockSet {
-				lock_id: [0u8; 8],
-				currency_id: DOT,
-				who: ALICE,
-				amount: 10,
-			}));
-
-			assert_ok!(<Tokens as MultiLockableCurrency<AccountId>>::remove_lock(
-				[0u8; 8], DOT, &ALICE
-			));
-			System::assert_last_event(Event::Tokens(crate::Event::LockRemoved {
-				lock_id: [0u8; 8],
-				currency_id: DOT,
-				who: ALICE,
-			}));
-		});
-}
-
-#[test]
-fn pallet_multi_reservable_currency_deposit_events() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 1000), (BOB, DOT, 1000)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(<Tokens as MultiReservableCurrency<AccountId>>::reserve(
-				DOT, &ALICE, 500
-			));
-			System::assert_last_event(Event::Tokens(crate::Event::Reserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 500,
-			}));
-
-			assert_eq!(
-				<Tokens as MultiReservableCurrency<AccountId>>::slash_reserved(DOT, &ALICE, 300),
-				0
-			);
-			System::assert_last_event(Event::Tokens(crate::Event::Slashed {
-				currency_id: DOT,
-				who: ALICE,
-				free_amount: 0,
-				reserved_amount: 300,
-			}));
-
-			assert_eq!(
-				<Tokens as MultiReservableCurrency<AccountId>>::unreserve(DOT, &ALICE, 100),
-				0
-			);
-			System::assert_last_event(Event::Tokens(crate::Event::Unreserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 100,
-			}));
-
-			assert_ok!(<Tokens as MultiReservableCurrency<AccountId>>::repatriate_reserved(
-				DOT,
-				&ALICE,
-				&BOB,
-				100,
-				BalanceStatus::Free
-			));
-			System::assert_last_event(Event::Tokens(crate::Event::ReserveRepatriated {
-				currency_id: DOT,
-				from: ALICE,
-				to: BOB,
-				amount: 100,
-				status: BalanceStatus::Free,
-			}));
-		});
-}
-
-#[test]
-fn pallet_fungibles_mutate_deposit_events() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(<Tokens as fungibles::Mutate<AccountId>>::mint_into(DOT, &ALICE, 500));
-			System::assert_last_event(Event::Tokens(crate::Event::Deposited {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 500,
-			}));
-			assert_ok!(<Tokens as fungibles::Mutate<AccountId>>::burn_from(DOT, &ALICE, 500));
-			System::assert_last_event(Event::Tokens(crate::Event::Withdrawn {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 500,
-			}));
-		});
-}
-
-#[test]
-fn pallet_fungibles_transfer_deposit_events() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(<Tokens as fungibles::Transfer<AccountId>>::transfer(
-				DOT, &ALICE, &BOB, 50, true
-			));
-			System::assert_last_event(Event::Tokens(crate::Event::Transfer {
-				currency_id: DOT,
-				from: ALICE,
-				to: BOB,
-				amount: 50,
-			}));
-		});
-}
-
-#[test]
-fn pallet_fungibles_unbalanced_deposit_events() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(<Tokens as MultiReservableCurrency<AccountId>>::reserve(DOT, &ALICE, 50));
-			assert_ok!(<Tokens as fungibles::Unbalanced<AccountId>>::set_balance(
-				DOT, &ALICE, 500
-			));
-			System::assert_last_event(Event::Tokens(crate::Event::BalanceSet {
-				currency_id: DOT,
-				who: ALICE,
-				free: 500,
-				reserved: 50,
-			}));
-
-			<Tokens as fungibles::Unbalanced<AccountId>>::set_total_issuance(DOT, 1000);
-			System::assert_last_event(Event::Tokens(crate::Event::TotalIssuanceSet {
-				currency_id: DOT,
-				amount: 1000,
-			}));
-		});
-}
-
-#[test]
-fn pallet_fungibles_mutate_hold_deposit_events() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			assert_ok!(<Tokens as fungibles::MutateHold<AccountId>>::hold(DOT, &ALICE, 50));
-			System::assert_last_event(Event::Tokens(crate::Event::Reserved {
-				currency_id: DOT,
-				who: ALICE,
-				amount: 50,
-			}));
-
-			assert_ok!(<Tokens as fungibles::MutateHold<AccountId>>::transfer_held(
-				DOT, &ALICE, &BOB, 50, true, true
-			));
-			System::assert_last_event(Event::Tokens(crate::Event::ReserveRepatriated {
-				currency_id: DOT,
-				from: ALICE,
-				to: BOB,
-				amount: 50,
-				status: BalanceStatus::Reserved,
-			}));
-			System::reset_events();
-			assert_eq!(
-				<Tokens as fungibles::MutateHold<AccountId>>::release(DOT, &BOB, 50, true),
-				Ok(50)
-			);
-			System::assert_last_event(Event::Tokens(crate::Event::Unreserved {
-				currency_id: DOT,
-				who: BOB,
-				amount: 50,
-			}));
-		});
-}
-
-#[test]
-fn currency_adapter_pallet_currency_deposit_events() {
-	ExtBuilder::default()
-		.balances(vec![(ALICE, DOT, 100), (BOB, DOT, 100)])
-		.build()
-		.execute_with(|| {
-			// Use std::mem::forget to get rid the returned imbalance.
-			std::mem::forget(<MockCurrencyAdapter as PalletCurrency<AccountId>>::burn(500));
-			System::assert_last_event(Event::Tokens(crate::Event::TotalIssuanceSet {
-				currency_id: DOT,
-				amount: 0,
-			}));
-
-			std::mem::forget(<MockCurrencyAdapter as PalletCurrency<AccountId>>::issue(200));
-			System::assert_last_event(Event::Tokens(crate::Event::TotalIssuanceSet {
-				currency_id: DOT,
-				amount: 200,
-			}));
-
-			assert_ok!(<MockCurrencyAdapter as PalletCurrency<AccountId>>::transfer(
-				&ALICE,
-				&BOB,
-				50,
-				ExistenceRequirement::AllowDeath
-			));
-			System::assert_last_event(Event::Tokens(crate::Event::Transfer {
-				currency_id: DOT,
-				from: ALICE,
-				to: BOB,
-				amount: 50,
-			}));
-
-			assert_ok!(<Tokens as MultiReservableCurrency<AccountId>>::reserve(DOT, &BOB, 50));
-			std::mem::forget(<MockCurrencyAdapter as PalletCurrency<AccountId>>::slash(&BOB, 110));
-			System::assert_last_event(Event::Tokens(crate::Event::Slashed {
-				currency_id: DOT,
-				who: BOB,
-				free_amount: 100,
-				reserved_amount: 10,
-			}));
-
-			std::mem::forget(<MockCurrencyAdapter as PalletCurrency<AccountId>>::make_free_balance_be(&BOB, 200));
-			System::assert_last_event(Event::Tokens(crate::Event::BalanceSet {
-				currency_id: DOT,
-				who: BOB,
-				free: 200,
-				reserved: 40,
-			}));
 		});
 }
