+++ tokens/src/tests_currency_adapter.rs
@@ -1,0 +1,640 @@
+//! Unit tests for the tokens module.
+
+#![cfg(test)]
+
+use super::*;
+use frame_support::{assert_noop, assert_ok};
+use mock::*;
+
+#[test]
+fn currency_adapter_ensure_currency_adapter_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			assert_eq!(Tokens::total_issuance(DOT), 100);
+			assert_eq!(Tokens::total_balance(DOT, &TREASURY_ACCOUNT), 100);
+			assert_eq!(Tokens::reserved_balance(DOT, &TREASURY_ACCOUNT), 0);
+			assert_eq!(Tokens::free_balance(DOT, &TREASURY_ACCOUNT), 100);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_balance(&TREASURY_ACCOUNT),
+				100
+			);
+			assert!(<Runtime as pallet_elections_phragmen::Config>::Currency::can_slash(
+				&TREASURY_ACCOUNT,
+				10
+			));
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
+				100
+			);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::minimum_balance(),
+				2
+			);
+			assert!(<Runtime as pallet_elections_phragmen::Config>::Currency::can_reserve(
+				&TREASURY_ACCOUNT,
+				5
+			));
+
+			// burn
+			let imbalance = <Runtime as pallet_elections_phragmen::Config>::Currency::burn(10);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
+				90
+			);
+			drop(imbalance);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
+				100
+			);
+
+			// issue
+			let imbalance = <Runtime as pallet_elections_phragmen::Config>::Currency::issue(20);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
+				120
+			);
+			drop(imbalance);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
+				100
+			);
+
+			// transfer
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
+				100
+			);
+			assert_ok!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::ensure_can_withdraw(
+					&TREASURY_ACCOUNT,
+					10,
+					WithdrawReasons::TRANSFER,
+					0
+				)
+			);
+			assert_ok!(<Runtime as pallet_elections_phragmen::Config>::Currency::transfer(
+				&TREASURY_ACCOUNT,
+				&ALICE,
+				11,
+				ExistenceRequirement::KeepAlive
+			));
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
+				89
+			);
+
+			// deposit
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
+				100
+			);
+			let imbalance = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 11);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
+				100
+			);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
+				100
+			);
+			drop(imbalance);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
+				100
+			);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
+				111
+			);
+
+			// withdraw
+			let imbalance = <Runtime as pallet_elections_phragmen::Config>::Currency::withdraw(
+				&TREASURY_ACCOUNT,
+				10,
+				WithdrawReasons::TRANSFER,
+				ExistenceRequirement::KeepAlive,
+			);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
+				90
+			);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
+				111
+			);
+			drop(imbalance);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::free_balance(&TREASURY_ACCOUNT),
+				90
+			);
+			assert_eq!(
+				<Runtime as pallet_elections_phragmen::Config>::Currency::total_issuance(),
+				101
+			);
+		});
+}
+
+#[test]
+fn currency_adapter_burn_must_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			let init_total_issuance = TreasuryCurrencyAdapter::total_issuance();
+			let imbalance = TreasuryCurrencyAdapter::burn(10);
+			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), init_total_issuance - 10);
+			drop(imbalance);
+			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), init_total_issuance);
+		});
+}
+
+#[test]
+fn currency_adapter_reserving_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
+
+		assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 111);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 111);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 0);
+
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 69));
+
+		assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 111);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 42);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 69);
+	});
+}
+
+#[test]
+fn currency_adapter_balance_transfer_when_reserved_should_not_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 69));
+		assert_noop!(
+			TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 69, ExistenceRequirement::AllowDeath),
+			Error::<Runtime>::BalanceTooLow,
+		);
+	});
+}
+
+#[test]
+fn currency_adapter_deducting_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 69));
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 42);
+	});
+}
+
+#[test]
+fn currency_adapter_refunding_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 42);
+		Tokens::set_reserved_balance(DOT, &TREASURY_ACCOUNT, 69);
+		TreasuryCurrencyAdapter::unreserve(&TREASURY_ACCOUNT, 69);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 111);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 0);
+	});
+}
+
+#[test]
+fn currency_adapter_slashing_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 69));
+		assert!(TreasuryCurrencyAdapter::slash(&TREASURY_ACCOUNT, 69).1.is_zero());
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 42);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 42);
+	});
+}
+
+#[test]
+fn currency_adapter_slashing_incomplete_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 42);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 42);
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 21));
+		assert_eq!(TreasuryCurrencyAdapter::slash(&TREASURY_ACCOUNT, 69).1, 27);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 0);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 0);
+	});
+}
+
+#[test]
+fn currency_adapter_basic_locking_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 100);
+			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 91, WithdrawReasons::all());
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 10, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+		});
+}
+
+#[test]
+fn currency_adapter_partial_locking_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 5, WithdrawReasons::all());
+			assert_ok!(TreasuryCurrencyAdapter::transfer(
+				&TREASURY_ACCOUNT,
+				&ALICE,
+				2,
+				ExistenceRequirement::AllowDeath
+			));
+		});
+}
+
+#[test]
+fn currency_adapter_lock_removal_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, u64::max_value(), WithdrawReasons::all());
+			TreasuryCurrencyAdapter::remove_lock(ID_1, &TREASURY_ACCOUNT);
+			assert_ok!(TreasuryCurrencyAdapter::transfer(
+				&TREASURY_ACCOUNT,
+				&ALICE,
+				2,
+				ExistenceRequirement::AllowDeath
+			));
+		});
+}
+
+#[test]
+fn currency_adapter_lock_replacement_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, u64::max_value(), WithdrawReasons::all());
+			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 5, WithdrawReasons::all());
+			assert_ok!(TreasuryCurrencyAdapter::transfer(
+				&TREASURY_ACCOUNT,
+				&ALICE,
+				2,
+				ExistenceRequirement::AllowDeath
+			));
+		});
+}
+
+#[test]
+fn currency_adapter_double_locking_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 5, WithdrawReasons::empty());
+			TreasuryCurrencyAdapter::set_lock(ID_2, &TREASURY_ACCOUNT, 5, WithdrawReasons::all());
+			assert_ok!(TreasuryCurrencyAdapter::transfer(
+				&TREASURY_ACCOUNT,
+				&ALICE,
+				2,
+				ExistenceRequirement::AllowDeath
+			));
+		});
+}
+
+#[test]
+fn currency_adapter_combination_locking_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			// withdrawReasons not work
+			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, u64::max_value(), WithdrawReasons::empty());
+			TreasuryCurrencyAdapter::set_lock(ID_2, &TREASURY_ACCOUNT, 0, WithdrawReasons::all());
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 2, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+		});
+}
+
+#[test]
+fn currency_adapter_lock_value_extension_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 100, WithdrawReasons::all());
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 6, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 2, WithdrawReasons::all());
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 6, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 8, WithdrawReasons::all());
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 3, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+		});
+}
+
+#[test]
+fn currency_adapter_lock_block_number_extension_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 200, WithdrawReasons::all());
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 6, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 90, WithdrawReasons::all());
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 6, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+			System::set_block_number(2);
+			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 90, WithdrawReasons::all());
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 3, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+		});
+}
+
+#[test]
+fn currency_adapter_lock_reasons_extension_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			TreasuryCurrencyAdapter::set_lock(ID_1, &TREASURY_ACCOUNT, 90, WithdrawReasons::TRANSFER);
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 11, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 90, WithdrawReasons::empty());
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 11, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+			TreasuryCurrencyAdapter::extend_lock(ID_1, &TREASURY_ACCOUNT, 90, WithdrawReasons::RESERVE);
+			assert_noop!(
+				TreasuryCurrencyAdapter::transfer(&TREASURY_ACCOUNT, &ALICE, 11, ExistenceRequirement::AllowDeath),
+				Error::<Runtime>::LiquidityRestrictions
+			);
+		});
+}
+
+#[test]
+fn currency_adapter_deposit_creating_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(ALICE, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 100);
+			assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 0);
+			let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 2);
+			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 102);
+			assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 2);
+
+			assert_eq!(TreasuryCurrencyAdapter::total_balance(&ALICE), 100);
+			let _ = TreasuryCurrencyAdapter::deposit_creating(&ALICE, 1);
+			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 103);
+			assert_eq!(TreasuryCurrencyAdapter::total_balance(&ALICE), 101);
+		});
+}
+
+#[test]
+fn currency_adapter_deposit_into_existing_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(ALICE, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 0);
+			assert_noop!(
+				TreasuryCurrencyAdapter::deposit_into_existing(&TREASURY_ACCOUNT, 10).map(drop),
+				Error::<Runtime>::DeadAccount,
+			);
+
+			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 100);
+			assert_eq!(TreasuryCurrencyAdapter::total_balance(&ALICE), 100);
+			assert_ok!(TreasuryCurrencyAdapter::deposit_into_existing(&ALICE, 10).map(drop));
+			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 110);
+			assert_eq!(TreasuryCurrencyAdapter::total_balance(&ALICE), 110);
+		});
+}
+
+#[test]
+fn currency_adapter_reward_should_work() {
+	ExtBuilder::default()
+		.balances(vec![(TREASURY_ACCOUNT, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 100);
+			assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 100);
+			assert_ok!(TreasuryCurrencyAdapter::deposit_into_existing(&TREASURY_ACCOUNT, 10).map(drop));
+			assert_eq!(TreasuryCurrencyAdapter::total_balance(&TREASURY_ACCOUNT), 110);
+			assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 110);
+		});
+}
+
+#[test]
+fn currency_adapter_slashing_reserved_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 111);
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 111));
+		assert_eq!(TreasuryCurrencyAdapter::slash_reserved(&TREASURY_ACCOUNT, 42).1, 0);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 69);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 69);
+	});
+}
+
+#[test]
+fn currency_adapter_slashing_incomplete_reserved_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 111);
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 42));
+		assert_eq!(TreasuryCurrencyAdapter::slash_reserved(&TREASURY_ACCOUNT, 69).1, 27);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 69);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 0);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 69);
+	});
+}
+
+#[test]
+fn currency_adapter_repatriating_reserved_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 110);
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&ALICE, 2);
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 110));
+		assert_ok!(
+			TreasuryCurrencyAdapter::repatriate_reserved(&TREASURY_ACCOUNT, &ALICE, 41, Status::Free),
+			0
+		);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 69);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&ALICE), 0);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&ALICE), 43);
+	});
+}
+
+#[test]
+fn currency_adapter_transferring_reserved_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 110);
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&ALICE, 2);
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 110));
+		assert_ok!(
+			TreasuryCurrencyAdapter::repatriate_reserved(&TREASURY_ACCOUNT, &ALICE, 41, Status::Reserved),
+			0
+		);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 69);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&ALICE), 41);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&ALICE), 2);
+	});
+}
+
+#[test]
+fn currency_adapter_transferring_reserved_balance_to_nonexistent_should_fail() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 111));
+		assert_ok!(TreasuryCurrencyAdapter::repatriate_reserved(
+			&TREASURY_ACCOUNT,
+			&ALICE,
+			42,
+			Status::Free
+		));
+	});
+}
+
+#[test]
+fn currency_adapter_transferring_incomplete_reserved_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 110);
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&ALICE, 2);
+		assert_ok!(TreasuryCurrencyAdapter::reserve(&TREASURY_ACCOUNT, 41));
+		assert_ok!(
+			TreasuryCurrencyAdapter::repatriate_reserved(&TREASURY_ACCOUNT, &ALICE, 69, Status::Free),
+			28
+		);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&TREASURY_ACCOUNT), 0);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 69);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance(&ALICE), 0);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&ALICE), 43);
+	});
+}
+
+#[test]
+fn currency_adapter_transferring_too_high_value_should_not_panic() {
+	ExtBuilder::default().build().execute_with(|| {
+		TreasuryCurrencyAdapter::make_free_balance_be(&TREASURY_ACCOUNT, u64::max_value());
+		TreasuryCurrencyAdapter::make_free_balance_be(&ALICE, 2);
+
+		assert_noop!(
+			TreasuryCurrencyAdapter::transfer(
+				&TREASURY_ACCOUNT,
+				&ALICE,
+				u64::max_value(),
+				ExistenceRequirement::AllowDeath
+			),
+			ArithmeticError::Overflow,
+		);
+
+		assert_eq!(
+			TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT),
+			u64::max_value()
+		);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&ALICE), 2);
+	});
+}
+
+#[test]
+fn exceeding_max_locks_should_fail() {
+	ExtBuilder::default()
+		.balances(vec![(ALICE, DOT, 100)])
+		.build()
+		.execute_with(|| {
+			assert_ok!(Tokens::set_lock(ID_1, DOT, &ALICE, 10));
+			assert_eq!(Tokens::locks(ALICE, DOT).len(), 1);
+			assert_ok!(Tokens::set_lock(ID_2, DOT, &ALICE, 10));
+			assert_eq!(Tokens::locks(ALICE, DOT).len(), 2);
+			assert_noop!(
+				Tokens::set_lock(ID_3, DOT, &ALICE, 10),
+				Error::<Runtime>::MaxLocksExceeded
+			);
+			assert_eq!(Tokens::locks(ALICE, DOT).len(), 2);
+		});
+}
+
+#[test]
+fn currency_adapter_slashing_named_reserved_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 111);
+		assert_ok!(TreasuryCurrencyAdapter::reserve_named(&RID_1, &TREASURY_ACCOUNT, 111));
+		assert_eq!(
+			TreasuryCurrencyAdapter::slash_reserved_named(&RID_1, &TREASURY_ACCOUNT, 42).1,
+			0
+		);
+		assert_eq!(
+			TreasuryCurrencyAdapter::reserved_balance_named(&RID_1, &TREASURY_ACCOUNT),
+			69
+		);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 69);
+	});
+}
+
+#[test]
+fn currency_adapter_named_slashing_incomplete_reserved_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 111);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 111);
+		assert_ok!(TreasuryCurrencyAdapter::reserve_named(&RID_1, &TREASURY_ACCOUNT, 42));
+		assert_eq!(
+			TreasuryCurrencyAdapter::slash_reserved_named(&RID_1, &TREASURY_ACCOUNT, 69).1,
+			27
+		);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 69);
+		assert_eq!(
+			TreasuryCurrencyAdapter::reserved_balance_named(&RID_1, &TREASURY_ACCOUNT),
+			0
+		);
+		assert_eq!(TreasuryCurrencyAdapter::total_issuance(), 69);
+	});
+}
+
+#[test]
+fn currency_adapter_repatriating_named_reserved_balance_should_work() {
+	ExtBuilder::default().build().execute_with(|| {
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&TREASURY_ACCOUNT, 110);
+		let _ = TreasuryCurrencyAdapter::deposit_creating(&ALICE, 2);
+		assert_ok!(TreasuryCurrencyAdapter::reserve_named(&RID_1, &TREASURY_ACCOUNT, 110));
+		assert_ok!(
+			TreasuryCurrencyAdapter::repatriate_reserved_named(&RID_1, &TREASURY_ACCOUNT, &ALICE, 41, Status::Free),
+			0
+		);
+		assert_eq!(
+			TreasuryCurrencyAdapter::reserved_balance_named(&RID_1, &TREASURY_ACCOUNT),
+			69
+		);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&TREASURY_ACCOUNT), 0);
+		assert_eq!(TreasuryCurrencyAdapter::reserved_balance_named(&RID_1, &ALICE), 0);
+		assert_eq!(TreasuryCurrencyAdapter::free_balance(&ALICE), 43);
+	});
+}
